% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_var_and_quintile}
\alias{plot_var_and_quintile}
\title{Plot variables within quintiles of another variable}
\usage{
plot_var_and_quintile(data, exposure, outcome)
}
\arguments{
\item{data}{Dataset to use}

\item{exposure}{Name of exposure variable}

\item{outcome}{Name of outcome variable}
}
\description{
Plot variables within quintiles of another variable
}
